+++ openwave/xperiments/1_readme_first/__init__.py
@@ -1,0 +1,4 @@
+"""
+1/ README FIRST
+==============================
+"""
