+++ openwave/xperiments/3_thermal_waves/__init__.py
@@ -1,0 +1,4 @@
+"""
+3/ THERMAL WAVES EXPERIMENT
+==============================
+"""
