+++ openwave/xperiments/4_level0_granule_based/__init__.py
@@ -1,5 +1,5 @@
 """
-LEVEL-0: Granule-Based Medium (educational tools)
+4/ LEVEL-0: Granule-Based Medium (educational tools)
 
 Experiments exploring wave dynamics using granule-based medium simulation.
 """
