+++ openwave/xperiments/5_level1_wave_field/__init__.py
@@ -1,5 +1,5 @@
 """
-LEVEL-1: Wave-Field Medium (analytical tools)
+5/ LEVEL-1: Wave-Field Medium (analytical tools)
 
 Experiments exploring wave dynamics using wave-field medium simulation.
 """
